pyproject.toml
[project]
name = "data-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.43.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-aiplatform>=1.127.0",
    "google-cloud-bigquery>=3.38.0",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "statsmodels>=0.14.5",
    "streamlit>=1.51.0",
]
